[gardening] Move ActorIsolation method definitions from Decl.cpp -> ActorIsolation.cpp.
Just noticed this while working in the general area.
